requirements.txt
# ==================== CORE WEB FRAMEWORK ====================
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# ==================== DATABASE ====================
pymongo==4.9.0
motor==3.3.2

# ==================== AUTHENTICATION & SECURITY ====================
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ==================== ENVIRONMENT & CONFIGURATION ====================
python-dotenv==1.0.1
pydantic==2.10.3

# ==================== HTTP CLIENT ====================
httpx==0.28.0

# ==================== TASK SCHEDULING ====================
apscheduler==3.10.4
pytz==2024.2

# ==================== DATA PROCESSING (if you use Excel imports) ====================
pandas==2.1.4
openpyxl==3.1.2